requirements.txt
fastapi
uvicorn
python-multipart
jinja2
redis
rq
psycopg2-binary
gunicorn